workspaceInvite route modified
