+++ src/admin/admin.controller.ts
@@ -3,7 +3,6 @@ import { CreateInviteDto } from './invite/dto/invite.dto';
 import { AdminService } from './admin.service';
 import { AuthorizeGuard } from 'src/guards/authorize.guard';
 import { AdminDecorator } from 'src/auth/decorators/admin.decorator';
-import { CreateWorkspaceInviteDto } from 'src/workspace/dtos/create-workspace-invite.dto';
 
 @Controller('admin')
 export class AdminController {
@@ -18,14 +17,6 @@ export class AdminController {
         @Body() inviteDto:CreateInviteDto
     ){
         return this.adminService.sendInviteLink(inviteDto)
-    }
-
-
-    @UseGuards(AuthorizeGuard)
-    @AdminDecorator()
-    @Post('workspace/invite')
-    async sendInvite(@Body() inviteDto:CreateWorkspaceInviteDto){
-        return await this.adminService.sendWorkspaceInvite(inviteDto);
     }
 
 }
