+++ src/admin/admin.service.ts
@@ -9,9 +9,6 @@ import { CreateInviteDto } from './invite/dto/invite.dto';
 import { UserService } from 'src/user/user.service';
 import { randomBytes } from 'crypto';
 import { MailService } from 'src/mail/mail.service';
-import { WorkspaceService } from 'src/workspace/workspace.service';
-import { WorkspaceInvite } from 'src/workspace/entity/workspace-invite.entity';
-import { CreateWorkspaceInviteDto } from 'src/workspace/dtos/create-workspace-invite.dto';
 
 @Injectable()
 export class AdminService {
@@ -27,9 +24,6 @@ export class AdminService {
 
         private readonly mailService: MailService,
 
-        @InjectRepository(WorkspaceInvite)
-                private workspaceInviteRepository: Repository<WorkspaceInvite>,
-                private workspaceService: WorkspaceService
     ) { }
 
     async createDefaultAdmin() {
@@ -106,51 +100,6 @@ export class AdminService {
         }
     }
 
-     async sendWorkspaceInvite(inviteDto: CreateWorkspaceInviteDto) {
-            try {
-                const { email, workspaceId } = inviteDto;
-    
-                const existingUser = await this.userService.findUserByMail(email)
-    
-                if (!existingUser) {
-                    throw new BadRequestException(`User with email ${email} does not exist`)
-                }
-    
-                const workspace = await this.workspaceService.findWorkspaceById(workspaceId)
-    
-                if (!workspace) {
-                    throw new BadRequestException(`Workspace not found`)
-                }
-    
-                const existingInvite = await this.workspaceInviteRepository.findOne({
-                    where: { email, workspaceId }
-                })
-                if (existingInvite && existingInvite.expiresAt > new Date()) {
-                    throw new BadRequestException(`Invite already sent to email ${email}`)
-                }
-    
-                const token = randomBytes(32).toString('hex')
-                const expiresAt = new Date(Date.now() + 24 * 60 * 60 * 1000)
-    
-                const invite = this.workspaceInviteRepository.create({
-                    email,
-                    token,
-                    workspaceId,
-                    isAccepted: false,
-                    expiresAt
-                })
-    
-                await this.workspaceInviteRepository.save(invite)
-                await this.mailService.sendWorkspaceInvite(email, token)
-    
-                return {
-                    success: true,
-                    message: `Workspace invitation mail sent successfully!`
-                }
-            } catch (error) {
-                throw error
-            }
-        }
     
 }
